+++ src/parser.rs
@@ -1,14 +1,13 @@
-///Imaging writing an entire parser for a Perl-like language just for a tiny little tool
-/// couldn't be me hahahahahahaha
 use winnow::{
-    IResult, Parser,
-    branch::alt,
-    bytes::complete::{tag, take, take_until},
-    character::complete::{alphanumeric1, char, digit1, space0},
-    combinator::{map, opt},
-    multi::many0,
-    sequence::{delimited, preceded},
+    Parser,
+    ascii::{alphanumeric1, digit1, space0},
+    combinator::{alt, delimited, opt, preceded, repeat_till},
+    error::{AddContext, ContextError, ParserError, StrContext},
+    token::{literal, take, take_until},
 };
+
+type WResult<O, E> = winnow::Result<O, E>;
+type Stream<'i> = &'i str;
 
 //For lack of a better name
 #[derive(Debug, PartialEq, Clone)]
@@ -32,102 +31,184 @@ pub enum TerminalCharacter {
     CowStart,
 }
 
-fn spaces_and_lines<'a>(input: &'a str) -> IResult<&'a str, TerminalCharacter> {
+fn spaces_and_lines<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
     alt((
-        map(tag("\r\n"), |_| TerminalCharacter::Newline),
-        map(tag("\n"), |_| TerminalCharacter::Newline),
-        map(tag(" "), |_| TerminalCharacter::Space),
+        literal("\r\n").map(|_| TerminalCharacter::Newline),
+        literal("\n").map(|_| TerminalCharacter::Newline),
+        literal(" ").map(|_| TerminalCharacter::Space),
     ))
-    .parse(input)
+    .parse_next(input)
 }
 
-fn misc_escapes<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(
-        delimited(tag("\\e["), alt((tag("39"), tag("49"))), tag("m")),
-        |esc: &str| match esc {
-            "39" => TerminalCharacter::DefaultForegroundColor,
-            "49" => TerminalCharacter::DefaultBackgroundColor,
-            _ => panic!(), //TODO Change this to some error handling
-        },
+fn misc_escapes<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    delimited(
+        literal("\\e["),
+        alt((literal("39"), literal("49"))),
+        literal("m"),
     )
-    .parse(i)
+    .map(|esc: &str| match esc {
+        "39" => TerminalCharacter::DefaultForegroundColor,
+        "49" => TerminalCharacter::DefaultBackgroundColor,
+        _ => panic!(), //TODO Change this to some error handling
+    })
+    .parse_next(input)
 }
 
-fn colors_256<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(
-        delimited(
-            tag("\\e["),
-            (
-                alt((tag("38"), tag("48"))),
-                tag(";"),
-                tag("5"),
-                tag(";"),
-                digit1,
-            ),
-            tag("m"),
+fn colors_256<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    delimited(
+        literal("\\e["),
+        (alt((literal("38"), literal("48"))), literal(";5;"), digit1),
+        literal("m"),
+    )
+    .map(|(color_type, _, color)| match color_type {
+        "38" => TerminalCharacter::TerminalForegroundColor256(str::parse(color).unwrap()),
+        "48" => TerminalCharacter::TerminalBackgroundColor256(str::parse(color).unwrap()),
+        _ => panic!(), //TODO change this to some kind of error handling
+    })
+    .parse_next(input)
+}
+
+fn truecolor<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    delimited(
+        literal("\\e["),
+        (
+            alt((literal("38"), literal("48"))),
+            literal(";2;"),
+            digit1,
+            literal(";"),
+            digit1,
+            literal(";"),
+            digit1,
         ),
-        |(color_type, _, _, _, color)| match color_type {
-            "38" => TerminalCharacter::TerminalForegroundColor256(str::parse(color).unwrap()),
-            "48" => TerminalCharacter::TerminalBackgroundColor256(str::parse(color).unwrap()),
+        literal("m"),
+    )
+    .map(|(color_type, _, red, _, green, _, blue)| {
+        match color_type {
+            "38" => TerminalCharacter::TerminalForegroundColorTruecolor(
+                str::parse(red).unwrap(),
+                str::parse(green).unwrap(),
+                str::parse(blue).unwrap(),
+            ),
+            "48" => TerminalCharacter::TerminalBackgroundColorTruecolor(
+                str::parse(red).unwrap(),
+                str::parse(green).unwrap(),
+                str::parse(blue).unwrap(),
+            ),
             _ => panic!(), //TODO change this to some kind of error handling
-        },
-    )
-    .parse(i)
+        }
+    })
+    .parse_next(input)
 }
 
-fn truecolor<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(
-        delimited(
-            tag("\\e["),
-            (
-                alt((tag("38"), tag("48"))),
-                tag(";"),
-                tag("2"),
-                tag(";"),
-                digit1,
-                tag(";"),
-                digit1,
-                tag(";"),
-                digit1,
-            ),
-            tag("m"),
+fn unicode_char<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    //Turns out, there are different ways to parse unicode escapes.
+    //This is my best attempt at covering them
+    alt((
+        //Xterm -\\N{U+xxxx}
+        delimited(literal("\\N{U+"), take(4 as usize), literal("}")).map(|code: &str| {
+            TerminalCharacter::UnicodeCharacter(
+                char::from_u32(u32::from_str_radix(code, 16).unwrap()).unwrap(),
+            )
+        }),
+        //ANSI - \\uxxxx
+        preceded(literal("\\u"), take(4 as usize)).map(|code: &str| {
+            TerminalCharacter::UnicodeCharacter(
+                char::from_u32(u32::from_str_radix(code, 16).unwrap()).unwrap(),
+            )
+        }),
+    ))
+    .parse_next(input)
+}
+
+//Fallback for a character that has an explicit escape
+fn escaped_char<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    preceded(literal("\\"), take(1usize))
+        .map(|character: &str| {
+            TerminalCharacter::EscapedUnicodeCharacter(character.chars().next().unwrap())
+        })
+        .parse_next(input)
+}
+
+fn comments<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    delimited(literal("#"), take_until(1.., "\n"), literal("\n"))
+        .map(|_| TerminalCharacter::Comment)
+        .parse_next(input)
+}
+
+///This parser is for random perl junk we see in files that we want to ignore since we aren't really a perl interpreter
+/// Some of it *is* useful when it comes to acting as a "barrier" between actual text we want to parse
+fn perl_junk<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    alt((
+        literal("binmode STDOUT, \":utf8\";\n"),
+        literal("binmode STDOUT, \":utf8\";\r\n"),
+    ))
+    .map(|_| TerminalCharacter::Comment)
+    .parse_next(input)
+}
+
+fn placeholders<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    alt((
+        literal("$thoughts").map(|_| TerminalCharacter::ThoughtPlaceholder),
+        literal("$eyes").map(|_| TerminalCharacter::EyePlaceholder),
+        literal("$tongue").map(|_| TerminalCharacter::TonguePlaceholder),
+    ))
+    .parse_next(input)
+}
+
+fn binding_name<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<&'a str, E> {
+    preceded(literal("$"), alphanumeric1).parse_next(input)
+}
+
+fn binding_value<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<Vec<TerminalCharacter>, E> {
+    delimited(
+        literal("\""),
+        repeat_till(
+            0..,
+            alt((
+                placeholders,
+                spaces_and_lines,
+                misc_escapes,
+                colors_256,
+                truecolor,
+                unicode_char,
+                escaped_char,
+                take(1 as usize).map(|c: &str| {
+                    TerminalCharacter::UnicodeCharacter(c.chars().into_iter().next().unwrap())
+                }),
+            )),
+            literal("\";"),
         ),
-        |(color_type, _, _, _, red, _, green, _, blue)| {
-            match color_type {
-                "38" => TerminalCharacter::TerminalForegroundColorTruecolor(
-                    str::parse(red).unwrap(),
-                    str::parse(green).unwrap(),
-                    str::parse(blue).unwrap(),
-                ),
-                "48" => TerminalCharacter::TerminalBackgroundColorTruecolor(
-                    str::parse(red).unwrap(),
-                    str::parse(green).unwrap(),
-                    str::parse(blue).unwrap(),
-                ),
-                _ => panic!(), //TODO change this to some kind of error handling
-            }
-        },
+        (literal("\";"), opt(alt((literal("\n"), literal("\r\n"))))),
     )
-    .parse(i)
+    .map(|(binding_val, _): (Vec<TerminalCharacter>, _)| binding_val)
+    .parse_next(input)
 }
 
-fn binding_name<'a>(i: &'a str) -> IResult<&'a str, &'a str> {
-    map(preceded(tag("$"), alphanumeric1), |name| name).parse(i)
-}
-
-fn binding_value<'a>(i: &'a str) -> IResult<&'a str, Vec<TerminalCharacter>> {
-    map(
-        delimited(
-            tag("\""),
-            many0(cow_string_tokens),
-            (tag("\";"), opt(alt((tag("\n"), tag("\r\n"))))),
-        ),
-        |binding_val| binding_val,
-    )
-    .parse(i)
-}
-
-fn var_binding<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
+fn var_binding<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
     //some assumptions of format we will be making here because perl has no BNF
     //1. Vars start with $, and are alphanumeric characters (at least one)
     //2. equals sign padded by any number of spaces on each side
@@ -135,155 +216,83 @@ fn var_binding<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
     // any number of characters up until an ending doublequote AND semicolon
     //4. If the line contianing the binding ends with a newline, take it since this isn't an actual part of the COWFILE
     //5. We are intentionally ignoring any bound variable calls in the binding value since it would literally be a pain in the ass to set up an "environment" for expanding the values
-    map(
-        (binding_name, space0, char('='), space0, binding_value),
-        |(binding_name, _, _, _, binding_value)| {
+    (binding_name, space0, literal("="), space0, binding_value)
+        .map(|(binding_name, _, _, _, binding_value)| {
             TerminalCharacter::VarBinding(binding_name.to_string(), binding_value)
-        },
-    )
-    .parse(i)
+        })
+        .parse_next(input)
 }
 
-fn bound_var_call<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(binding_name, |name| {
-        TerminalCharacter::BoundVarCall(name.to_string())
-    })
-    .parse(i)
+fn bound_var_call<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<TerminalCharacter, E> {
+    preceded(literal("$"), alphanumeric1)
+        .map(|name: &'a str| TerminalCharacter::BoundVarCall(name.to_string()))
+        .parse_next(input)
 }
 
-fn unicode_char<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    //Turns out, there are different ways to parse unicode escapes.
-    //This is my best attempt at covering them
-    alt((
-        //Xterm -\\N{U+xxxx}
-        map(
-            delimited(tag("\\N{U+"), take(4 as usize), tag("}")),
-            |code: &str| {
-                TerminalCharacter::UnicodeCharacter(
-                    char::from_u32(u32::from_str_radix(code, 16).unwrap()).unwrap(),
-                )
-            },
-        ),
-        //ANSI - \\uxxxx
-        map(preceded(tag("\\u"), take(4 as usize)), |code: &str| {
-            TerminalCharacter::UnicodeCharacter(
-                char::from_u32(u32::from_str_radix(code, 16).unwrap()).unwrap(),
-            )
-        }),
-    ))
-    .parse(i)
-}
-
-//Fallback for a character that has an explicit escape
-fn escaped_char<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(preceded(tag("\\"), take(1usize)), |character: &str| {
-        TerminalCharacter::EscapedUnicodeCharacter(character.chars().next().unwrap())
-    })
-    .parse(i)
-}
-
-fn comments<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    map(delimited(tag("#"), take_until("\n"), tag("\n")), |_| {
-        TerminalCharacter::Comment
-    })
-    .parse(i)
-}
-
-///This parser is for random perl junk we see in files that we want to ignore since we aren't really a perl interpreter
-/// Some of it *is* useful when it comes to acting as a "barrier" between actual text we want to parse
-fn perl_junk<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    // alt((
-    map(
-        alt((
-            // tag("EOC\n"),
-            // tag("EOC\r\n"),
-            // tag("\"@\n"),
-            // tag("\"@\r\n"),
-            tag("binmode STDOUT, \":utf8\";\n"),
-            tag("binmode STDOUT, \":utf8\";\r\n"),
-        )),
-        |_| TerminalCharacter::Comment,
-    )
-    // map(
-    //     alt((
-    //         tag("$the_cow = @\"\n"),
-    //         tag("$the_cow = @\"\r\n"),
-    //         tag("$the_cow =<<EOC;\n"),
-    //         tag("$the_cow =<<EOC;\r\n"),
-    //         tag("$the_cow = <<\"EOC\";\n"),
-    //         tag("$the_cow = <<\"EOC\";\r\n"),
-    //         tag("$the_cow = <<EOC;\n"),
-    //         tag("$the_cow = <<EOC;\r\n"),
-    //         tag("$the_cow = << EOC;\n"),
-    //         tag("$the_cow = << EOC;\r\n"),
-    //         tag("$the_cow = << EOC\n"),
-    //         tag("$the_cow = << EOC\r\n"),
-    //         tag("$the_cow = <<EOC\n"),
-    //         tag("$the_cow = <<EOC\r\n"),
-    //     )),
-    //     |_| TerminalCharacter::CowStart,
-    // ),
-    // ))
-    .parse(i)
-}
-
-fn placeholders<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    alt((
-        map(tag("$thoughts"), |_| TerminalCharacter::ThoughtPlaceholder),
-        map(tag("$eyes"), |_| TerminalCharacter::EyePlaceholder),
-        map(tag("$tongue"), |_| TerminalCharacter::TonguePlaceholder),
-    ))
-    .parse(i)
-}
-
-fn cow_string_tokens<'a>(i: &'a str) -> IResult<&'a str, TerminalCharacter> {
-    alt((
-        placeholders,
-        spaces_and_lines,
-        escaped_char,
-        misc_escapes,
-        colors_256,
-        truecolor,
-        unicode_char,
-        map(take(1 as usize), |c: &str| {
-            TerminalCharacter::UnicodeCharacter(c.chars().into_iter().next().unwrap())
-        }),
-    ))
-    .parse(i)
-}
-
-fn cow_string<'a>(i: &'a str) -> IResult<&'a str, Vec<TerminalCharacter>> {
+fn cow_string<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<Vec<TerminalCharacter>, E> {
     //NOTE this makes a flawed assumption where the perl delimiters don't have to match. But FWIW
     //it is not a significant bug honestly, most of these scripts _should_ be functional in OG perl
     let start = (
-        tag("$the_cow"),
+        literal("$the_cow"),
         space0,
-        tag("="),
+        literal("="),
         space0,
+        literal("<<"),
+        space0,
+        //NOTE This is easier than trying to form a bunch of sub parsers honestly
         alt((
-            map(
-                (
-                    tag("<<"),
-                    space0,
-                    alt((tag("\"EOC\""), tag("\"EOC\";"), tag("EOC"), tag("EOC;"))),
-                    alt((tag("\r\n"), tag("\n"))),
-                ),
-                |_| (),
-            ),
-            map(alt((tag("@\"\r\n"), tag("@\"\n"))), |_| ()),
+            literal("\"EOC\"\r\n"),
+            literal("\"EOC\"\n"),
+            literal("\"EOC\";\r\n"),
+            literal("\"EOC\";\n"),
+            literal("EOC\r\n"),
+            literal("EOC\n"),
+            literal("EOC;\r\n"),
+            literal("EOC;\n"),
+            literal("@\"\r\n"),
+            literal("@\"\n"),
         )),
     );
 
-    let end = (alt((tag("EOC"), tag("\"@"))), alt((tag("\r\n"), tag("\n"))));
-    delimited(start, many0(alt((bound_var_call, cow_string_tokens))), end)
-        .map(|mut chars| {
-            chars.insert(0usize, TerminalCharacter::CowStart);
-            chars
-        })
-        .parse(i)
+    preceded(
+        start,
+        repeat_till(
+            1..,
+            alt((
+                spaces_and_lines,
+                placeholders,
+                misc_escapes,
+                colors_256,
+                truecolor,
+                unicode_char,
+                escaped_char,
+                bound_var_call,
+                take(1 as usize).map(|c: &str| {
+                    TerminalCharacter::UnicodeCharacter(c.chars().into_iter().next().unwrap())
+                }),
+            )),
+            alt((
+                literal::<Stream<'a>, Stream<'a>, E>("EOC\r\n"),
+                literal::<Stream<'a>, Stream<'a>, E>("EOC\n"),
+                literal::<Stream<'a>, Stream<'a>, E>("\"@\r\n"),
+                literal::<Stream<'a>, Stream<'a>, E>("\"@\n"),
+            )),
+        ),
+    )
+    .map(|(mut chars, _): (Vec<TerminalCharacter>, _)| {
+        chars.insert(0usize, TerminalCharacter::CowStart);
+        chars
+    })
+    .parse_next(input)
 }
 
-pub fn cow_parser<'a>(input: &'a str) -> IResult<&'a str, Vec<TerminalCharacter>> {
+pub fn cow_parser<'a, E: ParserError<Stream<'a>> + AddContext<Stream<'a>, StrContext>>(
+    input: &mut Stream<'a>,
+) -> WResult<Vec<TerminalCharacter>, E> {
     alt((
         comments.map(|comment| vec![comment]),
         spaces_and_lines.map(|whitespace| vec![whitespace]),
@@ -291,5 +300,82 @@ pub fn cow_parser<'a>(input: &'a str) -> IResult<&'a str, Vec<TerminalCharacter>
         cow_string,
         var_binding.map(|binding| vec![binding]),
     ))
-    .parse(input)
+    // .parse(input)
+    .parse_next(input)
+}
+
+pub struct ParserIterator<'i> {
+    stream: Stream<'i>,
+    cow_started: bool,
+    prev_new_line: bool,
+    //NOTE Technically, better is to use generics, but we know what the internal iterator is
+    //composed of, so generic impl doesn't make too much sense
+    parsed_iter: Option<std::vec::IntoIter<TerminalCharacter>>,
+}
+
+///Based on the [std::iter::Flatten] implementation s
+impl<'i> ParserIterator<'i> {
+    pub fn new(input: &mut Stream<'i>) -> Self {
+        Self {
+            stream: &input,
+            cow_started: false,
+            prev_new_line: false,
+            parsed_iter: None,
+        }
+    }
+}
+
+impl<'i> Iterator for ParserIterator<'i> {
+    type Item = TerminalCharacter;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        match self.parsed_iter {
+            Some(ref mut iter) => match (iter).next() {
+                Some(val) => Some(val),
+                None => {
+                    self.parsed_iter = None;
+                    self.next() //NOTE this is obviously recursive. Should be a way to "fuse" iter here (effectively)
+                }
+            },
+            None => {
+                match cow_parser::<ContextError>.parse_next(&mut self.stream) {
+                    Ok(parsed) => {
+                        let iter = parsed.into_iter().filter(|parsed| match parsed {
+                            TerminalCharacter::Newline => {
+                                if self.prev_new_line {
+                                    false
+                                } else {
+                                    self.prev_new_line = self.cow_started;
+                                    true
+                                }
+                            }
+                            TerminalCharacter::Comment => true, //Should not alter newline state since it's not interpreted
+                            TerminalCharacter::CowStart => {
+                                self.cow_started = true;
+                                true
+                            }
+                            _ => {
+                                self.prev_new_line = false;
+                                true
+                            }
+                        });
+
+                        //HACK This is _really_ bad code
+                        //Basically since I can't really constrain the generic type of the filter predicate generic type
+                        // since closures can't be calculated that way and _this_ closure requires messing with this
+                        // iterator's internal state, I'm going to allocate a collected vector, then use _that vector's_
+                        // iterator as the reference iterator.
+                        //I can't think of a workaround. If you have a better solution, send patches and stop complaining
+                        // because I hate it too.
+                        let filtered: Vec<TerminalCharacter> = iter.collect();
+
+                        self.parsed_iter = Some(filtered.into_iter());
+
+                        self.next()
+                    }
+                    Err(_parse_err) => None,
+                }
+            }
+        }
+    }
 }
